Moved logic to observation results store
